% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_changes.R
\name{old_greensheets}
\alias{old_greensheets}
\title{Old greensheet categorization}
\usage{
old_greensheets(pend, plan)
}
\description{
Internal function used to re-create the old green sheet categorization
for comparison with the updated algorithm.
}
