% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rp_get_table}
\alias{rp_get_table}
\title{Retrieve RP table}
\usage{
rp_get_table(
  conn = NULL,
  tablename = c("v_crosstab_student_calc", "StudentInfoFields"),
  database = Sys.getenv("RP_DATABASE")
)
}
\arguments{
\item{conn}{DBI connection to use, see \code{rp_connect}.}

\item{tablename}{The name of the table to retrieve}

\item{database}{Name of the default database to connect to. Retrieved
from \code{RP_DATABASE} environment variable by default but can be overridden here,
see details.}
}
\description{
Retrieve RP table
}
\examples{
\dontshow{if (Sys.getenv("RP_ODBC_NAME")!=""&Sys.getenv("RP_DATABASE")!="") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
