% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_report.R
\name{rp_report_student_monitor}
\alias{rp_report_student_monitor}
\title{Create student monitor report}
\usage{
rp_report_student_monitor(conn = NULL, database = Sys.getenv("RP_DATABASE"))
}
\arguments{
\item{conn}{DBI connection to use, see \code{rp_connect}.}

\item{database}{Name of the default database to connect to. Retrieved
from \code{RP_DATABASE} environment variable by default but can be overridden here,
see details.}
}
\description{
Equivalent to \emph{Import/Export => Student => Monitors student file}
}
\examples{
\dontshow{if (Sys.getenv("RP_ODBC_NAME")!=""&Sys.getenv("RP_DATABASE")!="") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

student_monitor <- rp_report_student_monitor()

## data fields:
dplyr::glimpse(student_monitor[0, ])

## example pickup data:
dplyr::glimpse(dplyr::select(student_monitor, dplyr::starts_with("Pickup")))
\dontshow{\}) # examplesIf}
}
