% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_changes.R
\name{tripchange_report}
\alias{tripchange_report}
\title{Flag trips that have changed}
\usage{
tripchange_report(
  database_pend = Sys.getenv("RP_DATABASE_PEND"),
  database_plan = Sys.getenv("RP_DATABASE_PLAN"),
  load_threshold = 5,
  time_threshold = 3,
  distance_threshold = 100
)
}
\description{
Flag trips that have changed
}
