% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{otp_report}
\alias{otp_report}
\title{Generate a on-time performance report}
\usage{
otp_report(
  date = as.character(Sys.Date()),
  ampm = c("AM", "PM"),
  includeZonar = TRUE,
  cutoff_min = -25,
  cutoff_max = 60,
  rp_database = "SY_Live_RP",
  os_database = "Onscreen021919",
  rp_odbc_name = Sys.getenv("RP_ODBC_NAME"),
  uncoveredURL =
    "https://docs.google.com/spreadsheets/d/1OcpCGaPKZYaeazNuPjkzyWT8c-HgcFvMbzObuA-lZ50",
  mailTo = c("izahn@bostonpublicschools.org", "drosengard@bostonpublicschools.org",
    "ezanzerkia@bostonpublicschools.org"),
  mailFrom = "izahn@bostonpublicschools.org",
  TZ = "America/New_York",
  test = FALSE
)
}
\arguments{
\item{date}{Character vector of length 1 giving the date in YYY-MM-DD format.}

\item{ampm}{Character vector of length 1, either "AM", or "PM", use to
specify morning or evening OTP report.}

\item{includeZonar}{Logical vector of length 1. If \code{TRUE} zonar schedule data
will be retrieved and used for OTP calculations.}

\item{cutoff_min}{numeric vector of length one giving the minimum delay time in minutes
that will be considered a valid trip arrival. Usually this will be a negative value, e.g., -25 or so.}

\item{cutoff_max}{numeric vector of length one giving the maximum delay time in minutes
that will be considered a vlid trip arrival. Usuall this will be positive and relatively large, e.g., 60 or so.}

\item{rp_database}{Name of the RP database to connect to. Retrieved
from \code{RP_DATABASE} environment variable by default but can be overridden here,
see RVersatransRP package for details.}

\item{os_database}{Name of the Onscreen database to connect to.}

\item{rp_odbc_name}{Name of the Windows ODBC data source to use. Retrieved
from \code{RP_ODBC_NAME} environment variable by default but can be overridden
here, see RVersatransRP package for details.}

\item{uncoveredURL}{URL of a google sheet containing a list of uncovered trips.
Must contain columns named "Date", "AM/PM", "Route Set", and "Bus".}

\item{mailTo}{Character vector of email addresses to send notifications to.}

\item{mailFrom}{GMail address used to send notification emails.}

\item{TZ}{The timezone used by the database server.}

\item{test}{Logical vector of length one indicating whether to pull an abbreviated test data set.
Even with this the time needed to test this function will be long.}
}
\description{
Retrieves and combines data from Zonar and Versatrans and generates full OTP
data. The full schedule and OTP data is returned, it will need to be summarized
for public reporting.
}
\details{
TODO
}
\examples{
\dontshow{if (Sys.getenv("RP_ODBC_NAME")!=""&Sys.getenv("RP_DATABASE")!=""&Sys.getenv("ZONAR_PASS")!="") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

otp <- zonar_otp_report("2023-02-14", ampm = "AM", test = TRUE)
dplyr::glimpse(otp)
\dontshow{\}) # examplesIf}
}
